+++ pkg/models/site_stats.go
@@ -9,6 +9,6 @@ type SiteStats struct {
 	Pageviews   int64     `db:"pageviews"`
 	Sessions    int64     `db:"sessions"`
 	BounceRate  float64   `db:"bounce_rate"`
-	AvgDuration int64     `db:"avg_duration"`
+	AvgDuration float64   `db:"avg_duration"`
 	Date        time.Time `db:"date" json:"omitempty"`
 }
